# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "StudentPlanner.UI",
    "version": "1.0"
  },
  "paths": {
    "/academic-events": {
      "get": {
        "tags": [
          "AcademicEvents"
        ],
        "summary": "Retrieves all academic events.",
        "description": "This endpoint returns the complete list of academic events available\r\nin the system.",
        "responses": {
          "200": {
            "description": "Returns the list of academic events.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AcademicEventResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "AcademicEvents"
        ],
        "summary": "Creates a new academic event.",
        "requestBody": {
          "description": "The academic event to be created.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcademicEvent"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AcademicEvent"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns the created academic event.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicEventResponse"
                }
              }
            }
          },
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/academic-events/{id}": {
      "get": {
        "tags": [
          "AcademicEvents"
        ],
        "summary": "Retrieves a specific academic event by its identifier.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The unique identifier of the academic event.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the academic event.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicEventResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "404": {
            "description": "If the academic event does not exist."
          }
        }
      },
      "put": {
        "tags": [
          "AcademicEvents"
        ],
        "summary": "Updates an existing academic event.",
        "description": "The identifier in the route must match the identifier of the academic\r\nevent provided in the request body.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The unique identifier of the academic event.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The updated academic event data.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcademicEvent"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AcademicEvent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The academic event was updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicEventResponse"
                }
              }
            }
          },
          "400": {
            "description": "If the route identifier does not match the payload."
          },
          "500": {
            "description": "Internal Server Error"
          },
          "404": {
            "description": "If the academic event does not exist."
          }
        }
      },
      "delete": {
        "tags": [
          "AcademicEvents"
        ],
        "summary": "Deletes an existing academic event.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The unique identifier of the academic event.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The academic event was deleted successfully."
          },
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "404": {
            "description": "If the academic event does not exist."
          }
        }
      }
    },
    "/academic-events/{eventId}/subscribe": {
      "put": {
        "tags": [
          "AcademicEvents"
        ],
        "summary": "Subscribes the authenticated user to an academic event.",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "description": "Identifier of the academic event.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "If the user is not authorized."
          },
          "204": {
            "description": "Subscription created."
          },
          "404": {
            "description": "If the event does not exist."
          }
        }
      }
    },
    "/academic-events/{eventId}/unsubscribe": {
      "put": {
        "tags": [
          "AcademicEvents"
        ],
        "summary": "Unsubscribes the authenticated user from an academic event.",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "description": "Identifier of the academic event.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "If the user is not authorized."
          },
          "204": {
            "description": "Subscription removed."
          },
          "404": {
            "description": "If the event or subscription does not exist."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcademicEvent": {
        "required": [
          "details",
          "eventId",
          "facultyId"
        ],
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "details": {
            "$ref": "#/components/schemas/EventDetails"
          },
          "facultyId": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "faculty": {
            "$ref": "#/components/schemas/Faculty"
          }
        },
        "additionalProperties": false
      },
      "AcademicEventResponse": {
        "required": [
          "details",
          "eventId",
          "faculty"
        ],
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "faculty": {
            "$ref": "#/components/schemas/FacultyDTO"
          },
          "details": {
            "$ref": "#/components/schemas/EventDetailsOutputDTO"
          }
        },
        "additionalProperties": false
      },
      "ApplicationUser": {
        "required": [
          "firstName",
          "surname"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "normalizedUserName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "normalizedEmail": {
            "type": "string",
            "nullable": true
          },
          "emailConfirmed": {
            "type": "boolean"
          },
          "passwordHash": {
            "type": "string",
            "nullable": true
          },
          "securityStamp": {
            "type": "string",
            "nullable": true
          },
          "concurrencyStamp": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "phoneNumberConfirmed": {
            "type": "boolean"
          },
          "twoFactorEnabled": {
            "type": "boolean"
          },
          "lockoutEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lockoutEnabled": {
            "type": "boolean"
          },
          "accessFailedCount": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "surname": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "personalEvents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PersonalEvent"
            },
            "nullable": true
          },
          "eventRequests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventRequest"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EventDetails": {
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "maxLength": 70,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EventDetailsOutputDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EventRequest": {
        "required": [
          "facultyId",
          "requestId",
          "requestStatus",
          "requestType",
          "submissionDate",
          "userId"
        ],
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "requestType": {
            "$ref": "#/components/schemas/RequestType"
          },
          "requestStatus": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "submissionDate": {
            "type": "string",
            "format": "date-time"
          },
          "targetEventId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "details": {
            "$ref": "#/components/schemas/EventDetails"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "user": {
            "$ref": "#/components/schemas/ApplicationUser"
          },
          "facultyId": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "faculty": {
            "$ref": "#/components/schemas/Faculty"
          }
        },
        "additionalProperties": false
      },
...
      "Faculty": {
        "required": [
          "displayName",
          "facultyId",
          "facultyName"
        ],
        "type": "object",
        "properties": {
          "facultyId": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "facultyName": {
            "maxLength": 150,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "academicEvents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AcademicEvent"
            },
            "nullable": true
          },
          "eventRequests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventRequest"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FacultyDTO": {
        "required": [
          "displayName",
          "facultyId",
          "name"
        ],
        "type": "object",
        "properties": {
          "facultyId": {
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "displayName": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "PersonalEvent": {
        "required": [
          "details",
          "eventId",
          "userId"
        ],
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "details": {
            "$ref": "#/components/schemas/EventDetails"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "user": {
            "$ref": "#/components/schemas/ApplicationUser"
          }
        },
        "additionalProperties": false
      },
...
      "RequestStatus": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "RequestType": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
...
    }
  },
...
}